+++ src/SqlStatement.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phlib\Db;
+
+/**
+ * Interface for SQL statements, e.g. those that may be built dynamically, and rendered as a string.
+ * @private Not part of the BC promise. Placeholder for future implementation.
+ */
+interface SqlStatement extends \Stringable
+{
+}